Create a component for the hero section
